Changed uint64 back to int32

+++ auth/api/grpc/server.go
@@ -459,7 +459,7 @@ func decodeListSubjectsRequest(_ context.Context, grpcReq interface{}) (interfac
 		Relation:    req.GetRelation(),
 		Permission:  req.GetPermission(),
 		ObjectType:  req.GetObjectType(),
-		Object:      req.GetObject(), NextPageToken: req.GetNextPageToken(), Limit: req.GetLimit(),
+		Object:      req.GetObject(), NextPageToken: req.GetNextPageToken(), Limit: uint64(req.GetLimit()),
 	}, nil
 }
